v0.04 - Complete overhaul
Added a simple page loader
