+++ core/base classes/LayoutData.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: andre
+ * Date: 28-04-2018
+ * Time: 09:42
+ */
+
+
+class LayoutData {
+
+    var $template = "Home";
+    var $jumbotron = ["2","3"];
+    var $mainContent = ["1","2"];
+
+    function __construct($array)
+    {
+        if ($array != null) {
+            $this->template = $array["template"];
+            $this->jumbotron = $array["jumbotron"];
+            $this->mainContent = $array["mainContent"];
+        }
+    }
+
+}
