+++ layouts/home.json
@@ -1,0 +1,12 @@
+{
+	"template":"Home",
+	"jumbotron":
+		[
+			"banner"
+		],
+	"mainContent":
+		[
+			"homecontent",
+			"footer"
+		]
+}
