+++ modules/footer/index.php
@@ -1,0 +1,91 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: andre
+ * Date: 29-04-2018
+ * Time: 19:47
+ */
+
+?>
+<div class="footer">
+
+    <div class="container">
+
+        <div class="row">
+
+            <!-- Contact info -->
+            <div class="col-5 contactInformation">
+
+
+                <!-- Rows of information -->
+                <!-- Row 1 - Email adress -->
+                <div class="row info">
+                    <div class="col-2">
+                        <img class="contactImage" src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png">
+                    </div>
+                    <div class="col-10">
+                        <p class="contactInfo">
+                            Trafikskolen@kontant.dk
+                        </p>
+                    </div>
+                </div>
+                <!-- Row 1 - phone -->
+                <div class="row info">
+                    <div class="col-2">
+                        <img class="contactImage" src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png">
+                    </div>
+                    <div class="col-10">
+                        <p class="contactInfo">
+                            5131 5899
+                        </p>
+                    </div>
+                </div>
+                <!-- Row 1 - Location -->
+                <div class="row info">
+                    <div class="col-2">
+                        <img class="contactImage" src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png">
+                    </div>
+                    <div class="col-10">
+                        <p class="contactInfo">
+                            8800, Viborg
+                        </p>
+                    </div>
+                </div>
+                <!-- Row 1 - Abaout -->
+                <div class="row info">
+                    <div class="col-2">
+                        <img class="contactImage" src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png">
+                    </div>
+                    <div class="col-10">
+                        <p class="contactInfo">
+                            Mændende bag
+                        </p>
+                    </div>
+                </div>
+                <!-- Row 1 - Copyright -->
+                <div class="row info">
+                    <div class="col-2">
+                        <img class="contactImage" src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png">
+                    </div>
+                    <div class="col-10">
+                        <p class="contactInfo">
+                            Simon Eliasen & Nicklas Nyrup Lyngsøe
+                        </p>
+                    </div>
+                </div>
+
+
+            </div>
+
+            <div class="col-1 spacer">
+
+            </div>
+            <div class="col-6">
+
+            </div>
+
+        </div>
+
+    </div>
+
+</div>
