+++ templates/readme
@@ -1,0 +1,1 @@
+Only use lower-case for naming.
