--- views/about.php
@@ -1,1 +1,0 @@
-about
