--- views/about/index.php
@@ -1,1 +1,0 @@
-about
