--- views/home.php
@@ -1,2 +1,0 @@
-Home
-<a href="http://pearlnicklas.dk/index.php?page=about">link</a>
