--- views/home/content.php
@@ -1,83 +1,0 @@
-<div class="section">
-
-  <div class="container">
-
-    <div class="row">
-      <div class="col-3 element">
-        <img class="img-fluid" src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/19-256.png">
-        <h1> Statistik </h1>
-      </div>
-      <div class="col-3 element">
-        <img class="img-fluid" src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/63-256.png">
-        <h1> Glade kunder </h1>
-      </div>
-      <div class="col-3 element">
-        <img class="img-fluid" src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/88-256.png">
-        <h1> Hvem </h1>
-      </div>
-      <div class="col-3 element">
-        <img class="img-fluid" src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png">
-        <h1> Kontakt </h1>
-      </div>
-    </div>
-
-  </div>
-
-</div>
-<div class="section content">
-
-  <div class="container">
-
-    <div class="row">
-
-      <div class="col-12 headline">
-
-        <h1>Vi hjælper dig med at finde en kørerlærer i dit område!</h1>
-
-      </div>
-
-    </div>
-
-    <div class="row article">
-
-      <div class="col-8">
-
-        <p> Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque ultricies vitae odio et porttitor. Fusce quis risus pellentesque, commodo tellus ut, vestibulum libero. Morbi nec lectus nunc. Aenean orci diam, luctus tincidunt laoreet ac, vehicula in sapien. Mauris porttitor risus id suscipit consequat. Proin condimentum tortor in lectus hendrerit, in gravida sem interdum. Pellentesque interdum, enim sed commodo pellentesque, ligula augue lacinia lectus, eu scelerisque augue urna eget orci. Integer auctor mattis ligula at ultrices. Phasellus eget nulla sit amet lectus interdum pulvinar. Vestibulum rhoncus libero turpis, sit amet ultricies nisi consectetur in. Nam condimentum finibus tincidunt. Praesent id nibh ligula. Fusce ullamcorper rutrum eleifend. In consequat est erat, sagittis lobortis enim accumsan nec. Aenean venenatis ultrices nibh a tempus. Quisque molestie varius lacus, eu feugiat ex volutpat aliquet.</p>
-
-      </div>
-      <div class="col-1">
-
-      </div>
-
-      <div class="col-3">
-
-        <img class="image" src="http://media.safebee.com/assets/images/2015/7/elderly%20man%20thumbs%20up%20car.jpg.838x0_q67_crop-smart.jpg">
-
-      </div>
-
-    </div>
-    <div class="row article">
-
-      <div class="col-3">
-
-        <img class="image" src="http://media.safebee.com/assets/images/2015/7/elderly%20man%20thumbs%20up%20car.jpg.838x0_q67_crop-smart.jpg">
-
-      </div>
-      <div class="col-1">
-
-      </div>
-
-      <div class="col-8">
-
-        <p> Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque ultricies vitae odio et porttitor. Fusce quis risus pellentesque, commodo tellus ut, vestibulum libero. Morbi nec lectus nunc. Aenean orci diam, luctus tincidunt laoreet ac, vehicula in sapien. Mauris porttitor risus id suscipit consequat. Proin condimentum tortor in lectus hendrerit, in gravida sem interdum. Pellentesque interdum, enim sed commodo pellentesque, ligula augue lacinia lectus, eu scelerisque augue urna eget orci. Integer auctor mattis ligula at ultrices. Phasellus eget nulla sit amet lectus interdum pulvinar. Vestibulum rhoncus libero turpis, sit amet ultricies nisi consectetur in. Nam condimentum finibus tincidunt. Praesent id nibh ligula. Fusce ullamcorper rutrum eleifend. In consequat est erat, sagittis lobortis enim accumsan nec. Aenean venenatis ultrices nibh a tempus. Quisque molestie varius lacus, eu feugiat ex volutpat aliquet.</p>
-
-      </div>
-
-
-
-    </div>
-
-
-  </div>
-
-</div>
