--- views/home/index.php
@@ -1,12 +1,0 @@
-<?php
-// Including function that generates random image for header
-include "../functions/randomHeader.php";
-
-// Including header
-include "header.php";
-
-// Including content
-include "content.php";
-
-
-?>
