--- views/search/content.php
@@ -1,587 +1,0 @@
-<div class="content">
-  <div class="container results">
-    <!-- Displaying results -->
-    <div class="row result">
-      <div class="col-12">
-
-        <div class="row">
-
-          <!-- image for driver school -->
-          <div class="col-2 image">
-
-
-            <img class="img-fluid img" src="https://scontent-sea1-1.cdninstagram.com/t51.2885-19/11906133_107504862934838_1426164675_a.jpg" />
-
-
-          </div>
-
-
-          <div class="col-7">
-
-            <div class="row" style="">
-              <div class="col-12" style="">
-                <h1 class="headline"> Køreskole Lars Larsen</h1>
-              </div>
-            </div>
-
-            <div class="row">
-              <div class="col-12">
-                <p class="sub-text">
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent nisi felis, pretium eget lorem non, convallis dictum enim. Donec et enim vitae lacus hendrerit viverra ut id magna. Suspendisse posuere mi vitae gravida blandit. Aliquam placerat purus vitae tellus ullamcorper mollis. Vestibulum nec massa nulla. Cras malesuada, risus in rutrum volutpat, nulla metus dapibus ante, eu ornare elit ligula volutpat lacus. Aliquam mollis augue quam, in fermentum felis porttitor et. In non velit at turpis iaculis ullamcorper et sit amet magna. Quisque turpis eros, ultrices nec quam eu, commodo placerat eros. Suspendisse malesuada, risus eu vestibulum consequat, sapien ex consectetur eros, eu aliquam ligula orci ac tortor.</p>
-              </div>
-            </div>
-
-
-          </div>
-
-
-          <div class="col-3">
-            <div class="row" style="margin-top: 15px;">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/06-256.png" />
-              </div>
-              <div class="col-10 info">
-                42 56 32 75
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png" />
-              </div>
-              <div class="col-10 info">
-                nicklas.lyngsoee@gmail.com
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/94-256.png" />
-              </div>
-              <div class="col-10 info">
-                8800, Viborg
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/89-256.png" />
-              </div>
-              <div class="col-10 info">
-                4,7
-              </div>
-            </div>
-          </div>
-
-        </div>
-      </div>
-    </div>
-
-    <div class="row result">
-      <div class="col-12">
-
-        <div class="row">
-
-          <!-- image for driver school -->
-          <div class="col-2 image">
-
-
-            <img class="img-fluid img" src="http://www.njdriversafety.org/wp-content/uploads/defensive-driver-150x150.jpg" />
-
-
-          </div>
-
-
-          <div class="col-7">
-
-            <div class="row" style="">
-              <div class="col-12" style="">
-                <h1 class="headline"> Køreskole Søren Sørensen</h1>
-              </div>
-            </div>
-
-            <div class="row">
-              <div class="col-12">
-                <p class="sub-text">
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent nisi felis, pretium eget lorem non, convallis dictum enim. Donec et enim vitae lacus hendrerit viverra ut id magna. Suspendisse posuere mi vitae gravida blandit. Aliquam placerat purus vitae tellus ullamcorper mollis. Vestibulum nec massa nulla. Cras malesuada, risus in rutrum volutpat, nulla metus dapibus ante, eu ornare elit ligula volutpat lacus. Aliquam mollis augue quam, in fermentum felis porttitor et. In non velit at turpis iaculis ullamcorper et sit amet magna. Quisque turpis eros, ultrices nec quam eu, commodo placerat eros. Suspendisse malesuada, risus eu vestibulum consequat, sapien ex consectetur eros, eu aliquam ligula orci ac tortor.</p>
-              </div>
-            </div>
-
-
-          </div>
-
-
-          <div class="col-3">
-            <div class="row" style="margin-top: 15px;">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/06-256.png" />
-              </div>
-              <div class="col-10 info">
-                42 56 32 75
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png" />
-              </div>
-              <div class="col-10 info">
-                nicklas.lyngsoee@gmail.com
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/94-256.png" />
-              </div>
-              <div class="col-10 info">
-                8800, Viborg
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/89-256.png" />
-              </div>
-              <div class="col-10 info">
-                4,7
-              </div>
-            </div>
-          </div>
-
-        </div>
-      </div>
-    </div>
-    <div class="row result">
-      <div class="col-12">
-
-        <div class="row">
-
-          <!-- image for driver school -->
-          <div class="col-2 image">
-
-
-            <img class="img-fluid img" src="https://scontent-sea1-1.cdninstagram.com/t51.2885-19/11906133_107504862934838_1426164675_a.jpg" />
-
-
-          </div>
-
-
-          <div class="col-7">
-
-            <div class="row" style="">
-              <div class="col-12" style="">
-                <h1 class="headline"> Køreskole Lars Larsen</h1>
-              </div>
-            </div>
-
-            <div class="row">
-              <div class="col-12">
-                <p class="sub-text">
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent nisi felis, pretium eget lorem non, convallis dictum enim. Donec et enim vitae lacus hendrerit viverra ut id magna. Suspendisse posuere mi vitae gravida blandit. Aliquam placerat purus vitae tellus ullamcorper mollis. Vestibulum nec massa nulla. Cras malesuada, risus in rutrum volutpat, nulla metus dapibus ante, eu ornare elit ligula volutpat lacus. Aliquam mollis augue quam, in fermentum felis porttitor et. In non velit at turpis iaculis ullamcorper et sit amet magna. Quisque turpis eros, ultrices nec quam eu, commodo placerat eros. Suspendisse malesuada, risus eu vestibulum consequat, sapien ex consectetur eros, eu aliquam ligula orci ac tortor.</p>
-              </div>
-            </div>
-
-
-          </div>
-
-
-          <div class="col-3">
-            <div class="row" style="margin-top: 15px;">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/06-256.png" />
-              </div>
-              <div class="col-10 info">
-                42 56 32 75
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png" />
-              </div>
-              <div class="col-10 info">
-                nicklas.lyngsoee@gmail.com
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/94-256.png" />
-              </div>
-              <div class="col-10 info">
-                8800, Viborg
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/89-256.png" />
-              </div>
-              <div class="col-10 info">
-                4,7
-              </div>
-            </div>
-          </div>
-
-        </div>
-      </div>
-    </div>
-
-    <div class="row result">
-      <div class="col-12">
-
-        <div class="row">
-
-          <!-- image for driver school -->
-          <div class="col-2 image">
-
-
-            <img class="img-fluid img" src="http://www.njdriversafety.org/wp-content/uploads/defensive-driver-150x150.jpg" />
-
-
-          </div>
-
-
-          <div class="col-7">
-
-            <div class="row" style="">
-              <div class="col-12" style="">
-                <h1 class="headline"> Køreskole Søren Sørensen</h1>
-              </div>
-            </div>
-
-            <div class="row">
-              <div class="col-12">
-                <p class="sub-text">
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent nisi felis, pretium eget lorem non, convallis dictum enim. Donec et enim vitae lacus hendrerit viverra ut id magna. Suspendisse posuere mi vitae gravida blandit. Aliquam placerat purus vitae tellus ullamcorper mollis. Vestibulum nec massa nulla. Cras malesuada, risus in rutrum volutpat, nulla metus dapibus ante, eu ornare elit ligula volutpat lacus. Aliquam mollis augue quam, in fermentum felis porttitor et. In non velit at turpis iaculis ullamcorper et sit amet magna. Quisque turpis eros, ultrices nec quam eu, commodo placerat eros. Suspendisse malesuada, risus eu vestibulum consequat, sapien ex consectetur eros, eu aliquam ligula orci ac tortor.</p>
-              </div>
-            </div>
-
-
-          </div>
-
-
-          <div class="col-3">
-            <div class="row" style="margin-top: 15px;">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/06-256.png" />
-              </div>
-              <div class="col-10 info">
-                42 56 32 75
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png" />
-              </div>
-              <div class="col-10 info">
-                nicklas.lyngsoee@gmail.com
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/94-256.png" />
-              </div>
-              <div class="col-10 info">
-                8800, Viborg
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/89-256.png" />
-              </div>
-              <div class="col-10 info">
-                4,7
-              </div>
-            </div>
-          </div>
-
-        </div>
-      </div>
-    </div>
-    <div class="row result">
-      <div class="col-12">
-
-        <div class="row">
-
-          <!-- image for driver school -->
-          <div class="col-2 image">
-
-
-            <img class="img-fluid img" src="https://scontent-sea1-1.cdninstagram.com/t51.2885-19/11906133_107504862934838_1426164675_a.jpg" />
-
-
-          </div>
-
-
-          <div class="col-7">
-
-            <div class="row" style="">
-              <div class="col-12" style="">
-                <h1 class="headline"> Køreskole Lars Larsen</h1>
-              </div>
-            </div>
-
-            <div class="row">
-              <div class="col-12">
-                <p class="sub-text">
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent nisi felis, pretium eget lorem non, convallis dictum enim. Donec et enim vitae lacus hendrerit viverra ut id magna. Suspendisse posuere mi vitae gravida blandit. Aliquam placerat purus vitae tellus ullamcorper mollis. Vestibulum nec massa nulla. Cras malesuada, risus in rutrum volutpat, nulla metus dapibus ante, eu ornare elit ligula volutpat lacus. Aliquam mollis augue quam, in fermentum felis porttitor et. In non velit at turpis iaculis ullamcorper et sit amet magna. Quisque turpis eros, ultrices nec quam eu, commodo placerat eros. Suspendisse malesuada, risus eu vestibulum consequat, sapien ex consectetur eros, eu aliquam ligula orci ac tortor.</p>
-              </div>
-            </div>
-
-
-          </div>
-
-
-          <div class="col-3">
-            <div class="row" style="margin-top: 15px;">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/06-256.png" />
-              </div>
-              <div class="col-10 info">
-                42 56 32 75
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png" />
-              </div>
-              <div class="col-10 info">
-                nicklas.lyngsoee@gmail.com
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/94-256.png" />
-              </div>
-              <div class="col-10 info">
-                8800, Viborg
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/89-256.png" />
-              </div>
-              <div class="col-10 info">
-                4,7
-              </div>
-            </div>
-          </div>
-
-        </div>
-      </div>
-    </div>
-
-    <div class="row result">
-      <div class="col-12">
-
-        <div class="row">
-
-          <!-- image for driver school -->
-          <div class="col-2 image">
-
-
-            <img class="img-fluid img" src="http://www.njdriversafety.org/wp-content/uploads/defensive-driver-150x150.jpg" />
-
-
-          </div>
-
-
-          <div class="col-7">
-
-            <div class="row" style="">
-              <div class="col-12" style="">
-                <h1 class="headline"> Køreskole Søren Sørensen</h1>
-              </div>
-            </div>
-
-            <div class="row">
-              <div class="col-12">
-                <p class="sub-text">
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent nisi felis, pretium eget lorem non, convallis dictum enim. Donec et enim vitae lacus hendrerit viverra ut id magna. Suspendisse posuere mi vitae gravida blandit. Aliquam placerat purus vitae tellus ullamcorper mollis. Vestibulum nec massa nulla. Cras malesuada, risus in rutrum volutpat, nulla metus dapibus ante, eu ornare elit ligula volutpat lacus. Aliquam mollis augue quam, in fermentum felis porttitor et. In non velit at turpis iaculis ullamcorper et sit amet magna. Quisque turpis eros, ultrices nec quam eu, commodo placerat eros. Suspendisse malesuada, risus eu vestibulum consequat, sapien ex consectetur eros, eu aliquam ligula orci ac tortor.</p>
-              </div>
-            </div>
-
-
-          </div>
-
-
-          <div class="col-3">
-            <div class="row" style="margin-top: 15px;">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/06-256.png" />
-              </div>
-              <div class="col-10 info">
-                42 56 32 75
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png" />
-              </div>
-              <div class="col-10 info">
-                nicklas.lyngsoee@gmail.com
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/94-256.png" />
-              </div>
-              <div class="col-10 info">
-                8800, Viborg
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/89-256.png" />
-              </div>
-              <div class="col-10 info">
-                4,7
-              </div>
-            </div>
-          </div>
-
-        </div>
-      </div>
-    </div>
-    <div class="row result">
-      <div class="col-12">
-
-        <div class="row">
-
-          <!-- image for driver school -->
-          <div class="col-2 image">
-
-
-            <img class="img-fluid img" src="https://scontent-sea1-1.cdninstagram.com/t51.2885-19/11906133_107504862934838_1426164675_a.jpg" />
-
-
-          </div>
-
-
-          <div class="col-7">
-
-            <div class="row" style="">
-              <div class="col-12" style="">
-                <h1 class="headline"> Køreskole Lars Larsen</h1>
-              </div>
-            </div>
-
-            <div class="row">
-              <div class="col-12">
-                <p class="sub-text">
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent nisi felis, pretium eget lorem non, convallis dictum enim. Donec et enim vitae lacus hendrerit viverra ut id magna. Suspendisse posuere mi vitae gravida blandit. Aliquam placerat purus vitae tellus ullamcorper mollis. Vestibulum nec massa nulla. Cras malesuada, risus in rutrum volutpat, nulla metus dapibus ante, eu ornare elit ligula volutpat lacus. Aliquam mollis augue quam, in fermentum felis porttitor et. In non velit at turpis iaculis ullamcorper et sit amet magna. Quisque turpis eros, ultrices nec quam eu, commodo placerat eros. Suspendisse malesuada, risus eu vestibulum consequat, sapien ex consectetur eros, eu aliquam ligula orci ac tortor.</p>
-              </div>
-            </div>
-
-
-          </div>
-
-
-          <div class="col-3">
-            <div class="row" style="margin-top: 15px;">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/06-256.png" />
-              </div>
-              <div class="col-10 info">
-                42 56 32 75
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png" />
-              </div>
-              <div class="col-10 info">
-                nicklas.lyngsoee@gmail.com
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/94-256.png" />
-              </div>
-              <div class="col-10 info">
-                8800, Viborg
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/89-256.png" />
-              </div>
-              <div class="col-10 info">
-                4,7
-              </div>
-            </div>
-          </div>
-
-        </div>
-      </div>
-    </div>
-
-    <div class="row result">
-      <div class="col-12">
-
-        <div class="row">
-
-          <!-- image for driver school -->
-          <div class="col-2 image">
-
-
-            <img class="img-fluid img" src="http://www.njdriversafety.org/wp-content/uploads/defensive-driver-150x150.jpg" />
-
-
-          </div>
-
-
-          <div class="col-7">
-
-            <div class="row" style="">
-              <div class="col-12" style="">
-                <h1 class="headline"> Køreskole Søren Sørensen</h1>
-              </div>
-            </div>
-
-            <div class="row">
-              <div class="col-12">
-                <p class="sub-text">
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent nisi felis, pretium eget lorem non, convallis dictum enim. Donec et enim vitae lacus hendrerit viverra ut id magna. Suspendisse posuere mi vitae gravida blandit. Aliquam placerat purus vitae tellus ullamcorper mollis. Vestibulum nec massa nulla. Cras malesuada, risus in rutrum volutpat, nulla metus dapibus ante, eu ornare elit ligula volutpat lacus. Aliquam mollis augue quam, in fermentum felis porttitor et. In non velit at turpis iaculis ullamcorper et sit amet magna. Quisque turpis eros, ultrices nec quam eu, commodo placerat eros. Suspendisse malesuada, risus eu vestibulum consequat, sapien ex consectetur eros, eu aliquam ligula orci ac tortor.</p>
-              </div>
-            </div>
-
-
-          </div>
-
-
-          <div class="col-3">
-            <div class="row" style="margin-top: 15px;">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/06-256.png" />
-              </div>
-              <div class="col-10 info">
-                42 56 32 75
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/81-256.png" />
-              </div>
-              <div class="col-10 info">
-                nicklas.lyngsoee@gmail.com
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/94-256.png" />
-              </div>
-              <div class="col-10 info">
-                8800, Viborg
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-2 infoImg">
-                <img src="https://cdn2.iconfinder.com/data/icons/pittogrammi/142/89-256.png" />
-              </div>
-              <div class="col-10 info">
-                4,7
-              </div>
-            </div>
-          </div>
-
-        </div>
-      </div>
-    </div>
-
-  </div>
-
-</div>
