--- views/search/index.php
@@ -1,24 +1,0 @@
-
-<?php
-
-// Including search function to display search results
-include "../functions/search.php";
-?>
-
-
-
-<!--
-  Header section
-  Contains the stored search data and displays it.
-  Oppotunity to re-search with old data
--->
-<?php include "header.php"; ?>
-
-
-
-
-<!--
-  Content section
-  Contains results of the search
--->
-<?php include "content.php"; ?>
